gemini-code-context-engineering-template/use-cases/multi_agent_cli/requirements.txt
pydantic-ai
google-api-python-client
google-auth-oauthlib
python-dotenv
brave-search-api-client
